work in progress
Former-commit-id: d10fbf807211ecde72e9b61ad07187523d20f3ef
Former-commit-id: 4498b6c304026562bf78e09d280d5d82e72f6083

+++ src/wid_game_map_client.c
@@ -603,8 +603,8 @@ wid_game_map_client_replace_tile (widp w, double x, double y, thingp t)
 
     if (thing_is_explosion(t)) {
         wid_blit_scaling_to_pct_in(child, 1.0, 3.0, 500, 0);
-        dx = ((double)((rand() % 100) - 50)) / 100.0;
-        dy = ((double)((rand() % 100) - 50)) / 100.0;
+        dx = ((double)(((int)(rand() % 100)) - 50)) / 100.0;
+        dy = ((double)(((int)(rand() % 100)) - 50)) / 100.0;
         wid_fade_out(child, 500);
     }
